+++ Sources/CordialSDK/CordialAPI.swift
@@ -7,7 +7,6 @@
 //
 
 import UIKit
-import os.log
 
 @objc public class CordialAPI: NSObject {
     
@@ -83,7 +82,7 @@ import os.log
         
         let previousPrimaryKey = self.getContactPrimaryKey()
         
-        internalCordialAPI.setPreviousPrimaryKeyAndRemoveCurrent(previousPrimaryKey: previousPrimaryKey)
+        internalCordialAPI.setPreviousContactPrimaryKey(previousPrimaryKey: previousPrimaryKey)
         
         CoreDataManager.shared.deleteAllCoreDataByEntity(entityName: CoreDataManager.shared.contactLogoutRequest.entityName)
         
@@ -111,11 +110,9 @@ import os.log
             let sendContactLogoutRequest = SendContactLogoutRequest(primaryKey: previousPrimaryKey)
             ContactLogoutSender().sendContactLogout(sendContactLogoutRequest: sendContactLogoutRequest)
             
-            internalCordialAPI.setPreviousPrimaryKeyAndRemoveCurrent(previousPrimaryKey: previousPrimaryKey)
+            internalCordialAPI.setPreviousContactPrimaryKey(previousPrimaryKey: previousPrimaryKey)
         } else {
-            if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .error) {
-                os_log("Sending contact logout failed. Error: [User no login]", log: OSLog.cordialPushNotification, type: .error)
-            }
+            LoggerManager.shared.error(message: "Sending contact logout failed. Error: [User no login]", category: "CordialSDKPushNotification")
         }
     }
     
@@ -187,9 +184,7 @@ import os.log
                 CordialPushNotification.shared.registerForPushNotifications(options: options)
             }
         } else {
-            if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .info) {
-                os_log("Register for push notifications failed: pushesConfiguration not equals to SDK value", log: OSLog.cordialPushNotification, type: .info)
-            }
+            LoggerManager.shared.info(message: "Register for push notifications failed: pushesConfiguration not equals to SDK value", category: "CordialSDKPushNotification")
         }
     }