+++ Sources/CordialSDK/CoreData/Helper/InboxMessagesMarkReadUnreadCoreData.swift
@@ -8,7 +8,6 @@
 
 import Foundation
 import CoreData
-import os.log
 
 class InboxMessagesMarkReadUnreadCoreData {
     
@@ -28,9 +27,7 @@ class InboxMessagesMarkReadUnreadCoreData {
                 
                 try context.save()
             } catch let error {
-                if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .error) {
-                    os_log("CoreData Error: [%{public}@] Entity: [%{public}@]", log: OSLog.cordialCoreDataError, type: .error, error.localizedDescription, self.entityName)
-                }
+                LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
             }
         }
     }
@@ -53,18 +50,14 @@ class InboxMessagesMarkReadUnreadCoreData {
                 if let inboxMessagesMarkReadUnreadRequest = try NSKeyedUnarchiver.unarchiveTopLevelObjectWithData(data) as? InboxMessagesMarkReadUnreadRequest, !inboxMessagesMarkReadUnreadRequest.isError {
                     inboxMessagesMarkReadUnreadRequests.append(inboxMessagesMarkReadUnreadRequest)
                 } else {
-                    if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .error) {
-                        os_log("Failed unarchiving InboxMessagesMarkReadUnreadRequest", log: OSLog.cordialError, type: .error)
-                    }
+                    LoggerManager.shared.error(message: "Failed unarchiving InboxMessagesMarkReadUnreadRequest", category: "CordialSDKError")
                 }
                 
                 context.delete(managedObject)
                 try context.save()
             }
         } catch let error {
-            if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .error) {
-                os_log("CoreData Error: [%{public}@] Entity: [%{public}@]", log: OSLog.cordialCoreDataError, type: .error, error.localizedDescription, self.entityName)
-            }
+            LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
         }
 
         return inboxMessagesMarkReadUnreadRequests