+++ Sources/CordialSDK/Endpoints/InboxMessages/InboxMessagesGetter.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import os.log
 
 class InboxMessagesGetter {
     
@@ -18,15 +17,10 @@ class InboxMessagesGetter {
         if internalCordialAPI.isUserLogin() {
             if ReachabilityManager.shared.isConnectedToInternet {
                 if internalCordialAPI.getCurrentJWT() != nil {
-                    
-                    if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .info) {
-                        os_log("Fetching inbox messages", log: OSLog.cordialInboxMessages, type: .info)
-                    }
+                    LoggerManager.shared.info(message: "Fetching inbox messages", category: "CordialSDKInboxMessages")
                     
                     InboxMessages().getInboxMessages(pageRequest: pageRequest, inboxFilter: inboxFilter, contactKey: contactKey, onSuccess: { response in
-                        if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .info) {
-                            os_log("Inbox messages have been received successfully", log: OSLog.cordialInboxMessages, type: .info)
-                        }
+                        LoggerManager.shared.info(message: "Inbox messages have been received successfully", category: "CordialSDKInboxMessages")
                         
                         onSuccess(response)
                     }, onFailure: { error in