+++ Sources/CordialSDK/Endpoints/SendContactOrder/Helper/ContactOrdersSender.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import os.log
 
 class ContactOrdersSender {
     
@@ -21,20 +20,16 @@ class ContactOrdersSender {
             } else {
                 CoreDataManager.shared.contactOrderRequests.setContactOrderRequestsToCoreData(sendContactOrderRequests: sendContactOrderRequests)
                 
-                if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .info) {
-                    sendContactOrderRequests.forEach({ sendContactOrderRequest in
-                        os_log("Sending contact order failed. Saved to retry later. Request ID: [%{public}@] Error: [No Internet connection]", log: OSLog.cordialSendContactOrders, type: .info, sendContactOrderRequest.order.orderID)
-                    })
-                }
+                sendContactOrderRequests.forEach({ sendContactOrderRequest in
+                    LoggerManager.shared.info(message: "Sending contact order failed. Saved to retry later. Request ID: [\(sendContactOrderRequest.order.orderID)] Error: [No Internet connection]", category: "CordialSDKSendContactOrders")
+                })
             }
         } else {
             CoreDataManager.shared.contactOrderRequests.setContactOrderRequestsToCoreData(sendContactOrderRequests: sendContactOrderRequests)
             
-            if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .info) {
-                sendContactOrderRequests.forEach({ sendContactOrderRequest in
-                    os_log("Sending contact order failed. Saved to retry later. Request ID: [%{public}@] Error: [User no login]", log: OSLog.cordialSendContactOrders, type: .info, sendContactOrderRequest.order.orderID)
-                })
-            }
+            sendContactOrderRequests.forEach({ sendContactOrderRequest in
+                LoggerManager.shared.info(message: "Sending contact order failed. Saved to retry later. Request ID: [\(sendContactOrderRequest.order.orderID)] Error: [User no login]", category: "CordialSDKSendContactOrders")
+            })
         }
     }
     
@@ -42,18 +37,14 @@ class ContactOrdersSender {
         let sendContactOrders = SendContactOrders()
                 
         if InternalCordialAPI().getCurrentJWT() != nil {
-            
-            if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .info) {
-                sendContactOrderRequests.forEach({ sendContactOrderRequest in
-                    os_log("Sending contact order. Request ID: [%{public}@]", log: OSLog.cordialSendContactOrders, type: .info, sendContactOrderRequest.order.orderID)
-                    
-                    let payload = self.sendContactOrders.getSendContactOrderRequestJSON(sendContactOrderRequest: sendContactOrderRequest)
-                    os_log("Payload: %{public}@", log: OSLog.cordialSendContactOrders, type: .info, payload)
-                })
-            }
+            sendContactOrderRequests.forEach({ sendContactOrderRequest in
+                LoggerManager.shared.info(message: "Sending contact order. Request ID: [\(sendContactOrderRequest.order.orderID)]", category: "CordialSDKSendContactOrders")
+                
+                let payload = self.sendContactOrders.getSendContactOrderRequestJSON(sendContactOrderRequest: sendContactOrderRequest)
+                LoggerManager.shared.info(message: "Payload: \(payload)", category: "CordialSDKSendContactOrders")
+            })
             
             sendContactOrders.sendContactOrders(sendContactOrderRequests: sendContactOrderRequests)
-            
         } else {
             let responseError = ResponseError(message: "JWT is absent", statusCode: nil, responseBody: nil, systemError: nil)
             self.systemErrorHandler(sendContactOrderRequests: sendContactOrderRequests, error: responseError)
@@ -63,31 +54,25 @@ class ContactOrdersSender {
     }
     
     func completionHandler(sendContactOrderRequests: [SendContactOrderRequest]) {
-        if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .info) {
-            sendContactOrderRequests.forEach({ sendContactOrderRequest in
-                os_log("Order has been sent. Request ID: [%{public}@]", log: OSLog.cordialSendContactOrders, type: .info, sendContactOrderRequest.order.orderID)
-            })
-        }
+        sendContactOrderRequests.forEach({ sendContactOrderRequest in
+            LoggerManager.shared.info(message: "Order has been sent. Request ID: [\(sendContactOrderRequest.order.orderID)]", category: "CordialSDKSendContactOrders")
+        })
     }
     
     func systemErrorHandler(sendContactOrderRequests: [SendContactOrderRequest], error: ResponseError) {
         CoreDataManager.shared.contactOrderRequests.setContactOrderRequestsToCoreData(sendContactOrderRequests: sendContactOrderRequests)
         
-        if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .info) {
-            sendContactOrderRequests.forEach({ sendContactOrderRequest in
-                os_log("Sending contact order failed. Saved to retry later. Request ID: [%{public}@] Error: [%{public}@]", log: OSLog.cordialSendContactOrders, type: .info, sendContactOrderRequest.order.orderID, error.message)
-            })
-        }
+        sendContactOrderRequests.forEach({ sendContactOrderRequest in
+            LoggerManager.shared.info(message: "Sending contact order failed. Saved to retry later. Request ID: [\(sendContactOrderRequest.order.orderID)] Error: [\(error.message)]", category: "CordialSDKSendContactOrders")
+        })
     }
     
     func logicErrorHandler(sendContactOrderRequests: [SendContactOrderRequest], error: ResponseError) {
         NotificationCenter.default.post(name: .cordialSendContactOrdersLogicError, object: error)
         
-        if CordialApiConfiguration.shared.osLogManager.isAvailableOsLogLevelForPrint(osLogLevel: .error) {
-            sendContactOrderRequests.forEach({ sendContactOrderRequest in
-                os_log("Sending contact order failed. Will not retry. Request ID: [%{public}@] Error: [%{public}@]", log: OSLog.cordialSendContactOrders, type: .error, sendContactOrderRequest.order.orderID, error.message)
-            })
-        }
+        sendContactOrderRequests.forEach({ sendContactOrderRequest in
+            LoggerManager.shared.error(message: "Sending contact order failed. Will not retry. Request ID: [\(sendContactOrderRequest.order.orderID)] Error: [\(error.message)]", category: "CordialSDKSendContactOrders")
+        })
     }
     
 }